fix(editor): biome formatting
