+++ apps/editor/src/lib/components/scene/SceneCanvas.svelte
@@ -529,7 +529,7 @@
 
 <style>
   /* Soften Svelte Flow's default dotted background when no floor
-                               plan is set, but otherwise let its theming through. */
+                                 plan is set, but otherwise let its theming through. */
   :global(.svelte-flow__background) {
     background: #f8fafc;
   }
@@ -537,8 +537,8 @@
     stroke-linecap: round;
   }
   /* Make connection handles visible on hover so users can see where
-                             to drag from. Otherwise the fully-transparent handles leave the
-                             "how do I draw a wire" UX a guess. */
+                               to drag from. Otherwise the fully-transparent handles leave the
+                               "how do I draw a wire" UX a guess. */
   :global(.svelte-flow__node:hover .svelte-flow__handle) {
     /* biome-ignore lint/complexity/noImportantStyles: overrides Svelte Flow defaults */
     opacity: 1 !important;
@@ -548,15 +548,15 @@
     height: 8px;
   }
   /* Read-only cue: don't reveal connection handles on hover in view
-                   mode. Keep size + DOM presence so Svelte Flow can still resolve
-                   edge endpoint positions from each handle's bounding rect — only
-                   opacity is dropped. */
+                     mode. Keep size + DOM presence so Svelte Flow can still resolve
+                     edge endpoint positions from each handle's bounding rect — only
+                     opacity is dropped. */
   .scene-canvas-readonly :global(.svelte-flow__node:hover .svelte-flow__handle) {
     /* biome-ignore lint/complexity/noImportantStyles: overrides Svelte Flow defaults */
     opacity: 0 !important;
   }
   /* Placement-pending: crosshair cursor on the pane so users see
-                   "click somewhere to drop the item". */
+                     "click somewhere to drop the item". */
   .scene-canvas-placing :global(.svelte-flow__pane) {
     /* biome-ignore lint/complexity/noImportantStyles: overrides Svelte Flow defaults */
     cursor: crosshair !important;
